fix adding new files when already processing

+++ filepile.js
@@ -24,6 +24,7 @@ function filequeue(name, proc_fn) {
 
     // put an item into the queue and return the item id
     function queue(details) {
+        debug('pile %j', details);
         var id = uuid();
         var filename = path.join(basedir, id + '.json');
         fs.writeFileSync(filename, JSON.stringify(details));
@@ -42,11 +43,6 @@ function filequeue(name, proc_fn) {
         retries: 1000000000
     };
 
-    var processing = 0;
-
-    // while we are processing, we see other changes
-    // we will need to process those
-
     // lock so other processes don't try to access the files
     lockfile.lock(lockfilename, lockopts, function(err) {
         // cannot lock, assume that something else has the lockfile
@@ -64,9 +60,10 @@ function filequeue(name, proc_fn) {
         }, 1000);
     });
 
-    function files_changed() {
-        // already working through shits
-        if (processing++ > 0) {
+    var processing = false;
+
+    function files_changed(type, fname) {
+        if (fname === '.lock' || processing) {
             return;
         }
 
@@ -74,37 +71,38 @@ function filequeue(name, proc_fn) {
             return name.indexOf('.lock') < 0;
         });
 
-        (function next(err) {
+        if (files.length === 0) {
+            return;
+        }
+
+        processing = true;
+
+        (function next() {
             var file = files.shift();
             if (!file) {
-                return (--processing > 0) ? files_changed() : null;
+                processing = false;
+                return files_changed();
             }
 
             var filename = path.join(basedir, file);
-            single_file(filename, function(err) {
+            debug('processing file %s', filename);
+            var details = JSON.parse(fs.readFileSync(filename, 'utf8'));
+            proc_fn(details, function(err) {
                 if (err) {
                     var fail_filename = path.join(rejected_dir, file);
+                    debug('failed %s', fail_filename);
                     fs.renameSync(filename, fail_filename);
                     queue.emit('error', err);
+                }
+                else {
+                    debug('done processing %s', filename);
+                    fs.unlinkSync(filename);
                 }
 
                 next();
             });
         })();
     };
-
-    function single_file(filename, done) {
-        debug('processing file %s', filename);
-        var details = JSON.parse(fs.readFileSync(filename, 'utf8'));
-        proc_fn(details, function(err) {
-            if (err) {
-                return done(err);
-            }
-
-            fs.unlinkSync(filename);
-            done();
-        });
-    }
 
     // return a function we can use to queue
     return queue;
